+++ orby-app/src/components/orby-button/OrbyButton.js
@@ -1,0 +1,29 @@
+import React from "react";
+import { TouchableOpacity, Text, StyleSheet } from "react-native";
+import { COLORS } from "../../assets/colors";
+
+const OrbyButton = ({ onPress, title }) => {
+  return (
+    <TouchableOpacity style={styles.button} onPress={onPress}>
+      <Text style={styles.buttonText}>{title}</Text>
+    </TouchableOpacity>
+  );
+};
+
+const styles = StyleSheet.create({
+  button: {
+    backgroundColor: COLORS.primary,
+    padding: 12,
+    borderRadius: 25,
+    marginTop: 24,
+    marginBottom: 12,
+    alignItems: "center",
+  },
+  buttonText: {
+    color: COLORS.white,
+    fontWeight: "bold",
+    fontSize: 16,
+  },
+});
+
+export default OrbyButton;
